required dockerfile to run the wholeapp with docker-compose

+++ rest/src/main/java/com/ang/rest/domain/entities/TransactionDetailsId.java
@@ -10,6 +10,6 @@ import java.io.Serializable;
 @AllArgsConstructor
 @EqualsAndHashCode
 public class TransactionDetailsId implements Serializable {
-    private Integer transaction;
-    private Integer product;
+    private Long transaction;
+    private Long product;
 }
